+++ packages/core/src/agent/agent.test.ts
@@ -1,5 +1,4 @@
 import { randomUUID } from 'crypto';
-import { PassThrough } from 'stream';
 import { createOpenAI } from '@ai-sdk/openai';
 import { createOpenAI as createOpenAIV5 } from '@ai-sdk/openai-v5';
 import type { LanguageModelV2, LanguageModelV2TextPart } from '@ai-sdk/provider-v5';
@@ -8,7 +7,7 @@ import type { CoreMessage, LanguageModelV1, CoreSystemMessage } from 'ai';
 import { simulateReadableStream } from 'ai';
 import { MockLanguageModelV1 } from 'ai/test';
 import { stepCountIs } from 'ai-v5';
-import type { SystemModelMessage, UIMessageChunk } from 'ai-v5';
+import type { SystemModelMessage } from 'ai-v5';
 import { convertArrayToReadableStream, MockLanguageModelV2 } from 'ai-v5/test';
 import { config } from 'dotenv';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
@@ -17,16 +16,13 @@ import { z } from 'zod';
 import { TestIntegration } from '../integration/openapi-toolset.mock';
 import { noopLogger } from '../logger';
 import { Mastra } from '../mastra';
-import type { MastraMessageV1, MastraMessageV2, StorageThreadType } from '../memory';
+import type { MastraMessageV2, StorageThreadType } from '../memory';
 import { RuntimeContext } from '../runtime-context';
 import { createScorer } from '../scores';
 import { runScorer } from '../scores/hooks';
 import type { MastraModelOutput } from '../stream/base/output';
-import type { ChunkType } from '../stream/types';
-import { createMockModel } from '../test-utils/llm-mock';
 import { createTool } from '../tools';
 import { delay } from '../utils';
-import { CompositeVoice, MastraVoice } from '../voice';
 import { MessageList } from './message-list/index';
 import { assertNoDuplicateParts, MockMemory } from './test-utils';
 import { Agent } from './index';
@@ -2490,56 +2486,6 @@ function agentTests({ version }: { version: 'v1' | 'v2' }) {
       // Verify that default instructions were used when null was returned
       expect(capturedPrompt).toContain('you will generate a short title');
     });
-  });
-
-  describe(`${version} - agent llmPrompt`, () => {
-    it('should download assets from messages', async () => {
-      const agent = new Agent({
-        name: 'llmPrompt-agent',
-        instructions: 'test agent',
-        model: openaiModel,
-      });
-
-      let result;
-
-      if (version === 'v1') {
-        result = await agent.generateLegacy([
-          {
-            role: 'user',
-            content: [
-              {
-                type: 'image',
-                image: 'https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png',
-                mimeType: 'image/png',
-              },
-              {
-                type: 'text',
-                text: 'What is the photo?',
-              },
-            ],
-          },
-        ]);
-      } else {
-        result = await agent.generate([
-          {
-            role: 'user',
-            content: [
-              {
-                type: 'image',
-                image: 'https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png',
-                mimeType: 'image/png',
-              },
-              {
-                type: 'text',
-                text: 'What is the photo?',
-              },
-            ],
-          },
-        ]);
-      }
-
-      expect(result.text.toLowerCase()).toContain('google');
-    }, 10000);
   });
 
   describe(`${version} - agent tool handling`, () => {
@@ -9233,123 +9179,6 @@ function agentTests({ version }: { version: 'v1' | 'v2' }) {
 }
 
 describe('Agent Tests', () => {
-  describe('voice capabilities', () => {
-    class MockVoice extends MastraVoice {
-      async speak(): Promise<NodeJS.ReadableStream> {
-        const stream = new PassThrough();
-        stream.end('mock audio');
-        return stream;
-      }
-
-      async listen(): Promise<string> {
-        return 'mock transcription';
-      }
-
-      async getSpeakers() {
-        return [{ voiceId: 'mock-voice' }];
-      }
-    }
-
-    let voiceAgent: Agent;
-    beforeEach(() => {
-      voiceAgent = new Agent({
-        name: 'Voice Agent',
-        instructions: 'You are an agent with voice capabilities',
-        model: openai_v5('gpt-4o'),
-        voice: new CompositeVoice({
-          output: new MockVoice({
-            speaker: 'mock-voice',
-          }),
-          input: new MockVoice({
-            speaker: 'mock-voice',
-          }),
-        }),
-      });
-    });
-
-    describe('getSpeakers', () => {
-      it('should list available voices', async () => {
-        const speakers = await voiceAgent.voice?.getSpeakers();
-        expect(speakers).toEqual([{ voiceId: 'mock-voice' }]);
-      });
-    });
-
-    describe('speak', () => {
-      it('should generate audio stream from text', async () => {
-        const audioStream = await voiceAgent.voice?.speak('Hello World', {
-          speaker: 'mock-voice',
-        });
-
-        if (!audioStream) {
-          expect(audioStream).toBeDefined();
-          return;
-        }
-
-        const chunks: Buffer[] = [];
-        for await (const chunk of audioStream) {
-          chunks.push(Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk));
-        }
-        const audioBuffer = Buffer.concat(chunks);
-
-        expect(audioBuffer.toString()).toBe('mock audio');
-      });
-
-      it('should work with different parameters', async () => {
-        const audioStream = await voiceAgent.voice?.speak('Test with parameters', {
-          speaker: 'mock-voice',
-          speed: 0.5,
-        });
-
-        if (!audioStream) {
-          expect(audioStream).toBeDefined();
-          return;
-        }
-
-        const chunks: Buffer[] = [];
-        for await (const chunk of audioStream) {
-          chunks.push(Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk));
-        }
-        const audioBuffer = Buffer.concat(chunks);
-
-        expect(audioBuffer.toString()).toBe('mock audio');
-      });
-    });
-
-    describe('listen', () => {
-      it('should transcribe audio', async () => {
-        const audioStream = new PassThrough();
-        audioStream.end('test audio data');
-
-        const text = await voiceAgent.voice?.listen(audioStream);
-        expect(text).toBe('mock transcription');
-      });
-
-      it('should accept options', async () => {
-        const audioStream = new PassThrough();
-        audioStream.end('test audio data');
-
-        const text = await voiceAgent.voice?.listen(audioStream, {
-          language: 'en',
-        });
-        expect(text).toBe('mock transcription');
-      });
-    });
-
-    describe('error handling', () => {
-      it('should throw error when no voice provider is configured', async () => {
-        const agentWithoutVoice = new Agent({
-          name: 'No Voice Agent',
-          instructions: 'You are an agent without voice capabilities',
-          model: openai_v5('gpt-4o'),
-        });
-
-        await expect(agentWithoutVoice.voice.getSpeakers()).rejects.toThrow('No voice provider configured');
-        await expect(agentWithoutVoice.voice.speak('Test')).rejects.toThrow('No voice provider configured');
-        await expect(agentWithoutVoice.voice.listen(new PassThrough())).rejects.toThrow('No voice provider configured');
-      });
-    });
-  });
-
   it('should preserve empty assistant messages after tool use', () => {
     const messageList = new MessageList();
 
@@ -9484,202 +9313,6 @@ describe('Agent Tests', () => {
 
   agentTests({ version: 'v1' });
   agentTests({ version: 'v2' });
-
-  describe('agent.stopWhen', () => {
-    const weatherTool = createTool({
-      id: 'weather-tool',
-      description: 'Get weather for a location',
-      inputSchema: z.object({
-        location: z.string(),
-      }),
-      execute: async context => {
-        const { location } = context;
-        return {
-          temperature: 70,
-          feelsLike: 65,
-          humidity: 50,
-          windSpeed: 10,
-          windGust: 15,
-          conditions: 'sunny',
-          location,
-        };
-      },
-    });
-
-    const planActivities = createTool({
-      id: 'plan-activities',
-      description: 'Plan activities based on the weather',
-      inputSchema: z.object({
-        temperature: z.string(),
-      }),
-      execute: async () => {
-        return { activities: 'Plan activities based on the weather' };
-      },
-    });
-
-    const agent = new Agent({
-      name: 'test-step-boundaries',
-      instructions:
-        'You are a helpful assistant. Figure out the weather and then using that weather plan some activities. Always use the weather tool first, and then the plan activities tool with the result of the weather tool',
-      model: openai_v5('gpt-4o-mini'),
-      tools: {
-        weatherTool,
-        planActivities,
-      },
-    });
-
-    it('should demonstrate that stopWhen gets the proper step results in the first step', async () => {
-      let stopWhenCallCount = 0;
-      const stopWhenCalls: { callNumber: number; steps: any[] }[] = [];
-
-      const stream = await agent.stream('What should i be doing in Toronto today?', {
-        stopWhen: ({ steps }) => {
-          stopWhenCallCount++;
-          // Store the call details
-          stopWhenCalls.push({
-            callNumber: stopWhenCallCount,
-            steps: JSON.parse(JSON.stringify(steps)), // Deep copy to preserve state
-          });
-
-          // Check if any step has tool calls
-          const hasToolCalls = steps.some(step => {
-            return step.content && step.content.some(item => item.type === 'tool-call' || item.type === 'tool-result');
-          });
-
-          if (hasToolCalls) {
-            console.log(`Found tool calls in steps, attempting to stop...`);
-            // Try to stop immediately after finding tool calls
-            return true;
-          }
-
-          return false;
-        },
-      });
-
-      const chunks: ChunkType[] = [];
-      let stepStartCount = 0;
-      let foundToolCallChunk = false;
-      let foundTextChunk = false;
-      for await (const chunk of stream.fullStream) {
-        chunks.push(chunk);
-
-        if (chunk.type === 'step-start') {
-          stepStartCount++;
-        } else if (chunk.type === 'tool-call') {
-          foundToolCallChunk = true;
-        } else if (chunk.type === 'text-delta' && chunk.payload.text.trim()) {
-          foundTextChunk = true;
-        }
-      }
-
-      expect(foundToolCallChunk).toBe(true);
-
-      expect(stepStartCount).toBe(1);
-      expect(foundTextChunk).toBe(false);
-
-      expect(stopWhenCallCount).toBe(1);
-    }, 10000);
-
-    it('should not call stopWhen on the final step', async () => {
-      let stopWhenCallCount = 0;
-      const stopWhenCalls: { callNumber: number; stepCount: number }[] = [];
-
-      const trackStopWhenCalls = ({ steps }: { steps: any[] }) => {
-        stopWhenCallCount++;
-        stopWhenCalls.push({
-          callNumber: stopWhenCallCount,
-          stepCount: steps.length,
-        });
-
-        return false;
-      };
-
-      const stream = await agent.stream('What should i be doing in Toronto today?', {
-        stopWhen: trackStopWhenCalls,
-      });
-
-      let stepStartCount = 0;
-      let stepFinishCount = 0;
-
-      for await (const chunk of stream.fullStream) {
-        if (chunk.type === 'step-start') {
-          stepStartCount++;
-          console.log(`Step ${stepStartCount} started`);
-        } else if (chunk.type === 'step-finish') {
-          stepFinishCount++;
-          console.log(`Step ${stepFinishCount} finished`);
-        }
-      }
-
-      await stream.consumeStream();
-
-      // Verify that stopWhen is called n-1 times for n steps
-      expect(stepStartCount).toBe(stepFinishCount);
-      expect(stepStartCount).toBeGreaterThanOrEqual(2);
-
-      expect(stopWhenCallCount).toBe(stepStartCount - 1);
-
-      stopWhenCalls.forEach((call, index) => {
-        expect(call.stepCount).toBe(index + 1);
-      });
-    }, 25000);
-
-    it('should contain the correct content in the step results for both stopWhen and stream.steps', async () => {
-      const stopWhenContent: any[] = [];
-      const stream = await agent.stream('What should i be doing in Toronto today?', {
-        stopWhen: ({ steps }) => {
-          stopWhenContent.push(steps.at(-1)?.content);
-          return false;
-        },
-      });
-
-      await stream.consumeStream();
-
-      const steps = await stream.steps;
-
-      expect(stopWhenContent[0].length).toBe(2);
-      expect(stopWhenContent[1].length).toBe(2);
-
-      expect(steps[0].content.length).toBe(2);
-      expect(steps[1].content.length).toBe(2);
-      expect(steps[2].content.length).toBe(1);
-
-      expect(stopWhenContent[1]).not.toEqual(stopWhenContent[0]);
-    }, 20000);
-
-    it('should contain the correct content in the step results for both stopWhen and stream.steps with text and tool calls in the same step', async () => {
-      const agent = new Agent({
-        name: 'test-step-boundaries',
-        instructions:
-          'You are a helpful assistant. Figure out the weather and then using that weather plan some activities. Always use the weather tool first, and then the plan activities tool with the result of the weather tool. Every tool call you make IMMEDIATELY explain the tool results after executing the tool, before moving on to other steps or tool calls',
-        model: openai_v5('gpt-4o-mini'),
-        tools: {
-          weatherTool,
-          planActivities,
-        },
-      });
-
-      const stopWhenContent: any[] = [];
-      const stream = await agent.stream('What should i be doing in Toronto today?', {
-        stopWhen: ({ steps }) => {
-          stopWhenContent.push(steps.at(-1)?.content);
-          return false;
-        },
-      });
-
-      await stream.consumeStream();
-      const steps = await stream.steps;
-
-      expect(stopWhenContent[0].length).toBe(2);
-      expect(stopWhenContent[1].length).toBe(3); // text explaining the previous tool-call-results, tool-call, tool-result
-
-      expect(steps[0].content.length).toBe(2);
-      expect(steps[1].content.length).toBe(3); // text explaining the previous tool-call-results, tool-call, tool-result
-      expect(steps[2].content.length).toBe(1);
-
-      expect(stopWhenContent[1]).not.toEqual(stopWhenContent[0]);
-    }, 20000);
-  });
 });
 
 //     it('should accept and execute both Mastra and Vercel tools in Agent constructor', async () => {
@@ -9723,696 +9356,4 @@ describe('Agent Tests', () => {
 //       expect(mastraExecute).toHaveBeenCalled();
 //       expect(vercelExecute).toHaveBeenCalled();
 //     });
-
 // });
-
-describe('Stream ID Consistency', () => {
-  /**
-   * Test to verify that stream response IDs match database-saved message IDs
-   */
-
-  let memory: MockMemory;
-  let mastra: Mastra;
-
-  beforeEach(() => {
-    memory = new MockMemory();
-    mastra = new Mastra();
-  });
-
-  it('should return stream response IDs that can fetch saved messages from database', async () => {
-    const model = new MockLanguageModelV1({
-      doStream: async () => ({
-        stream: simulateReadableStream({
-          initialDelayInMs: 0,
-          chunkDelayInMs: 1,
-          chunks: [
-            { type: 'text-delta', textDelta: 'Hello! ' },
-            { type: 'text-delta', textDelta: 'I am ' },
-            { type: 'text-delta', textDelta: 'a helpful assistant.' },
-            {
-              type: 'finish',
-              finishReason: 'stop',
-              usage: { promptTokens: 10, completionTokens: 20 },
-            },
-          ],
-        }),
-        rawCall: { rawPrompt: [], rawSettings: {} },
-      }),
-    });
-
-    const agent = new Agent({
-      name: 'test-agent',
-      instructions: 'You are a helpful assistant.',
-      model,
-      // model: openai('gpt-4o'),
-      memory,
-    });
-
-    agent.__registerMastra(mastra);
-
-    const threadId = randomUUID();
-    const resourceId = 'test-resource';
-
-    const streamResult = await agent.streamLegacy('Hello!', {
-      threadId,
-      resourceId,
-    });
-
-    let streamResponseId: string | undefined;
-    for await (const _chunk of streamResult.fullStream) {
-      console.log('DEBUG chunk', _chunk);
-    }
-    await streamResult.consumeStream();
-
-    const finishedResult = streamResult;
-    const response = await finishedResult.response;
-
-    streamResponseId = response?.messages?.[0]?.id;
-
-    console.log('DEBUG streamResponseId', streamResponseId);
-    expect(streamResponseId).toBeDefined();
-
-    const savedMessages = await memory.getMessages({ threadId });
-
-    const messageById = savedMessages.find(m => m.id === streamResponseId);
-
-    expect(messageById).toBeDefined();
-    expect(messageById!.id).toBe(streamResponseId);
-  });
-
-  it('should use custom ID generator for streaming and keep stream response IDs consistent with database', async () => {
-    let customIdCounter = 0;
-    const customIdGenerator = vi.fn(() => `custom-id-${++customIdCounter}`);
-
-    const model = new MockLanguageModelV1({
-      doGenerate: async () => ({
-        rawCall: { rawPrompt: null, rawSettings: {} },
-        finishReason: 'stop' as const,
-        usage: { promptTokens: 10, completionTokens: 20 },
-        text: 'Hello! I am a helpful assistant.',
-      }),
-      doStream: async () => ({
-        stream: simulateReadableStream({
-          initialDelayInMs: 0,
-          chunkDelayInMs: 1,
-          chunks: [
-            { type: 'text-delta', textDelta: 'Hello! ' },
-            { type: 'text-delta', textDelta: 'I am ' },
-            { type: 'text-delta', textDelta: 'a helpful assistant.' },
-            {
-              type: 'finish',
-              finishReason: 'stop',
-              usage: { promptTokens: 10, completionTokens: 20 },
-            },
-          ],
-        }),
-        rawCall: { rawPrompt: [], rawSettings: {} },
-      }),
-    });
-
-    const mastraWithCustomId = new Mastra({
-      idGenerator: customIdGenerator,
-      logger: false,
-    });
-
-    const agent = new Agent({
-      name: 'test-agent',
-      instructions: 'You are a helpful assistant.',
-      model,
-      memory,
-    });
-
-    agent.__registerMastra(mastraWithCustomId);
-
-    const threadId = randomUUID();
-    const resourceId = 'test-resource';
-
-    const stream = await agent.streamLegacy('Hello!', { threadId, resourceId });
-
-    await stream.consumeStream();
-    const res = await stream.response;
-    const messageId = res.messages[0].id;
-
-    const savedMessages = await memory.getMessages({ threadId, selectBy: { include: [{ id: messageId }] } });
-
-    expect(savedMessages).toHaveLength(1);
-    expect(savedMessages[0].id).toBe(messageId);
-    expect(customIdGenerator).toHaveBeenCalled();
-  });
-
-  it('should return stream response IDs that can fetch saved messages from database', async () => {
-    const model = new MockLanguageModelV2({
-      doStream: async () => ({
-        rawCall: { rawPrompt: null, rawSettings: {} },
-        warnings: [],
-        stream: convertArrayToReadableStream([
-          {
-            type: 'stream-start',
-            warnings: [],
-          },
-          {
-            type: 'response-metadata',
-            id: 'v2-msg-xyz123',
-            modelId: 'mock-model-id',
-            timestamp: new Date(0),
-          },
-          { type: 'text-start', id: '1' },
-          { type: 'text-delta', id: '1', delta: 'Hello! ' },
-          { type: 'text-delta', id: '1', delta: 'I am a ' },
-          { type: 'text-delta', id: '1', delta: 'helpful assistant.' },
-          { type: 'text-end', id: '1' },
-          {
-            type: 'finish',
-            finishReason: 'stop',
-            usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
-          },
-        ]),
-      }),
-    });
-
-    const agent = new Agent({
-      name: 'test-agent',
-      instructions: 'You are a helpful assistant.',
-      model,
-      memory,
-    });
-
-    agent.__registerMastra(mastra);
-
-    const threadId = randomUUID();
-    const resourceId = 'test-resource';
-
-    const streamResult = await agent.stream('Hello!', {
-      threadId,
-      resourceId,
-    });
-
-    await streamResult.consumeStream();
-
-    let streamResponseId: string | undefined;
-    const res = await streamResult.response;
-    streamResponseId = res.uiMessages[0].id;
-
-    expect(streamResponseId).toBeDefined();
-
-    const savedMessages = await memory.getMessages({ threadId, selectBy: { include: [{ id: streamResponseId! }] } });
-    const messageById = savedMessages.find(m => m.id === streamResponseId);
-
-    expect(messageById).toBeDefined();
-    expect(messageById!.id).toBe(streamResponseId);
-  });
-
-  it('should use custom ID generator for stream and keep stream response IDs consistent with database', async () => {
-    let customIdCounter = 0;
-    const customIdGenerator = vi.fn(() => `custom-v2-id-${++customIdCounter}`);
-
-    const model = new MockLanguageModelV2({
-      doGenerate: async () => ({
-        rawCall: { rawPrompt: null, rawSettings: {} },
-        finishReason: 'stop',
-        usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
-        content: [
-          {
-            type: 'text',
-            text: 'Hello! I am a helpful assistant.',
-          },
-        ],
-        warnings: [],
-      }),
-      doStream: async () => ({
-        rawCall: { rawPrompt: null, rawSettings: {} },
-        warnings: [],
-        stream: convertArrayToReadableStream([
-          {
-            type: 'stream-start',
-            warnings: [],
-          },
-          {
-            type: 'response-metadata',
-            id: 'custom-v2-msg-xyz123',
-            modelId: 'mock-model-id',
-            timestamp: new Date(0),
-          },
-          { type: 'text-start', id: '1' },
-          { type: 'text-delta', id: '1', delta: 'Hello! ' },
-          { type: 'text-delta', id: '1', delta: 'I am a ' },
-          { type: 'text-delta', id: '1', delta: 'helpful assistant.' },
-          { type: 'text-end', id: '1' },
-          {
-            type: 'finish',
-            finishReason: 'stop',
-            usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
-          },
-        ]),
-      }),
-    });
-
-    const mastraWithCustomId = new Mastra({
-      idGenerator: customIdGenerator,
-      logger: false,
-    });
-
-    const agent = new Agent({
-      name: 'test-agent',
-      instructions: 'You are a helpful assistant.',
-      model,
-      memory,
-    });
-
-    agent.__registerMastra(mastraWithCustomId);
-
-    const threadId = randomUUID();
-    const resourceId = 'test-resource';
-
-    const stream = await agent.stream('Hello!', { threadId, resourceId });
-
-    await stream.consumeStream();
-    const res = await stream.response;
-    const messageId = res.uiMessages[0].id;
-    const savedMessages = await memory.getMessages({ threadId, selectBy: { include: [{ id: messageId }] } });
-    expect(savedMessages).toHaveLength(1);
-    expect(savedMessages[0].id).toBe(messageId);
-    expect(customIdGenerator).toHaveBeenCalled();
-  });
-
-  describe('onFinish callback with structured output', () => {
-    it('should include object field in onFinish callback when using structured output', async () => {
-      const mockModel = new MockLanguageModelV2({
-        doStream: async () => ({
-          rawCall: { rawPrompt: null, rawSettings: {} },
-          stream: convertArrayToReadableStream([
-            { type: 'stream-start', warnings: [] },
-            { type: 'response-metadata', id: 'id-0', modelId: 'mock-model-id', timestamp: new Date(0) },
-            { type: 'text-start', id: '1' },
-            { type: 'text-delta', id: '1', delta: '{"name":"John","age":30}' },
-            { type: 'text-end', id: '1' },
-            {
-              type: 'finish',
-              finishReason: 'stop',
-              usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
-            },
-          ]),
-        }),
-      });
-
-      const agent = new Agent({
-        name: 'test-structured-output-onfinish',
-        instructions: 'You are a helpful assistant.',
-        model: mockModel,
-      });
-
-      let onFinishResult: any = null;
-      let onFinishCalled = false;
-
-      const outputSchema = z.object({
-        name: z.string(),
-        age: z.number(),
-      });
-
-      const response = await agent.generate(
-        [
-          {
-            role: 'user',
-            content: 'Extract the person data',
-          },
-        ],
-        {
-          output: outputSchema,
-          onFinish: async result => {
-            onFinishCalled = true;
-            onFinishResult = result;
-          },
-        },
-      );
-
-      // Wait a bit to ensure onFinish is called
-      await new Promise(resolve => setTimeout(resolve, 100));
-
-      // The main function should return the structured data correctly
-      expect(response.object).toBeDefined();
-      expect(response.object).toEqual({ name: 'John', age: 30 });
-
-      // onFinish should have been called
-      expect(onFinishCalled).toBe(true);
-      expect(onFinishResult).toBeDefined();
-
-      // The fix: onFinish result should now include the object field
-      expect(onFinishResult.object).toBeDefined();
-      expect(onFinishResult.object).toEqual({ name: 'John', age: 30 });
-    }, 10000); // Increase timeout to 10 seconds
-  });
-
-  it('should include object field in onFinish callback when using structuredOutput key', async () => {
-    const mockModel = new MockLanguageModelV2({
-      doStream: async () => ({
-        rawCall: { rawPrompt: null, rawSettings: {} },
-        stream: convertArrayToReadableStream([
-          { type: 'stream-start', warnings: [] },
-          { type: 'response-metadata', id: 'id-0', modelId: 'mock-model-id', timestamp: new Date(0) },
-          { type: 'text-start', id: '1' },
-          { type: 'text-delta', id: '1', delta: 'The person is John who is 30 years old' },
-          { type: 'text-end', id: '1' },
-          {
-            type: 'finish',
-            finishReason: 'stop',
-            usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
-          },
-        ]),
-      }),
-    });
-
-    const structuringModel = new MockLanguageModelV2({
-      doStream: async () => ({
-        rawCall: { rawPrompt: null, rawSettings: {} },
-        stream: convertArrayToReadableStream([
-          { type: 'stream-start', warnings: [] },
-          { type: 'response-metadata', id: 'id-0', modelId: 'mock-model-id', timestamp: new Date(0) },
-          { type: 'text-start', id: '1' },
-          { type: 'text-delta', id: '1', delta: '{"name":"John","age":30}' },
-          { type: 'text-end', id: '1' },
-          {
-            type: 'finish',
-            finishReason: 'stop',
-            usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
-          },
-        ]),
-      }),
-    });
-
-    const agent = new Agent({
-      name: 'test-structured-output-processor-onfinish',
-      instructions: 'You are a helpful assistant.',
-      model: mockModel,
-    });
-
-    let onFinishResult: any = null;
-    let onFinishCalled = false;
-
-    const outputSchema = z.object({
-      name: z.string(),
-      age: z.number(),
-    });
-
-    const response = await agent.generate(
-      [
-        {
-          role: 'user',
-          content: 'Extract the person data',
-        },
-      ],
-      {
-        structuredOutput: {
-          schema: outputSchema,
-          model: structuringModel,
-        },
-        onFinish: async result => {
-          onFinishCalled = true;
-          onFinishResult = result;
-        },
-      },
-    );
-
-    // Wait a bit to ensure onFinish is called
-    await new Promise(resolve => setTimeout(resolve, 100));
-
-    // The main function should return the structured data correctly
-    expect(response.object).toBeDefined();
-    expect(response.object).toEqual({ name: 'John', age: 30 });
-
-    // onFinish should have been called
-    expect(onFinishCalled).toBe(true);
-    expect(onFinishResult).toBeDefined();
-
-    // The fix: onFinish result should now include the object field
-    expect(onFinishResult.object).toBeDefined();
-    expect(onFinishResult.object).toEqual({ name: 'John', age: 30 });
-  }, 10000); // Increase timeout to 10 seconds
-
-  it('should have messageIds when using toUIMessageStream', async () => {
-    const mockMemory = new MockMemory();
-    const threadId = randomUUID();
-    const resourceId = 'user-1';
-    const initialThread: StorageThreadType = {
-      id: threadId,
-      resourceId,
-      metadata: { client: 'initial' },
-      createdAt: new Date(),
-      updatedAt: new Date(),
-    };
-    await mockMemory.saveThread({ thread: initialThread });
-
-    const agent = new Agent({
-      name: 'test-agent',
-      instructions: 'You are a helpful assistant.',
-      model: openai_v5('gpt-4o'),
-      memory: mockMemory,
-    });
-
-    const mastra = new Mastra({ agents: { test: agent } });
-    agent.__registerMastra(mastra);
-
-    const stream = await agent.stream('Hello!', { threadId, resourceId, format: 'aisdk' });
-
-    // Get the UI message stream
-    const uiStream = stream.toUIMessageStream();
-
-    // Collect all chunks from the UI stream to verify message ID is present
-    const chunks: UIMessageChunk[] = [];
-    const reader = uiStream.getReader();
-
-    try {
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) break;
-        chunks.push(value);
-      }
-    } finally {
-      reader.releaseLock();
-    }
-
-    const messages = await mockMemory.getMessages({ threadId });
-    console.log('messages', messages);
-
-    const assistantMessage = messages.find((m: MastraMessageV1) => m.role === 'assistant');
-    console.log('assistantMessage', assistantMessage);
-    const startEvent = chunks.find(chunk => chunk.type === 'start');
-    console.log('startEvent', startEvent);
-
-    expect(assistantMessage?.id).toBe(startEvent?.messageId);
-  });
-});
-
-describe('Agent usage tracking', () => {
-  describe('Agent usage tracking (VNext paths)', () => {
-    describe('generate', () => {
-      it('should expose usage with inputTokens and outputTokens (AI SDK v5 format)', async () => {
-        // Create a V2 mock that returns usage in AI SDK v5 format
-        const model = new MockLanguageModelV2({
-          doGenerate: async () => ({
-            content: [{ type: 'text', text: 'Hello world!' }],
-            finishReason: 'stop',
-            usage: {
-              inputTokens: 10,
-              outputTokens: 20,
-              totalTokens: 30,
-            },
-            warnings: [],
-          }),
-          doStream: async () => {
-            return {
-              stream: convertArrayToReadableStream([
-                { type: 'text-start', id: '1' },
-                { type: 'text-delta', id: '1', delta: 'Hello world!' },
-                { type: 'text-end', id: '1' },
-                {
-                  type: 'finish',
-                  finishReason: 'stop',
-                  usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
-                },
-              ]),
-            };
-          },
-        });
-
-        const agent = new Agent({
-          name: 'test-agent',
-          model,
-          instructions: 'You are a helpful assistant',
-        });
-
-        const result = await agent.generate('Hello');
-
-        // Check that usage exists
-        expect(result.usage).toBeDefined();
-        console.log('generate usage:', result.usage);
-
-        // Check v5 format keys
-        expect(result.usage.inputTokens).toBe(10);
-        expect(result.usage.outputTokens).toBe(20);
-        expect(result.usage.totalTokens).toBe(30);
-
-        // Ensure backward compatibility keys are NOT present
-        expect((result.usage as any).promptTokens).toBeUndefined();
-        expect((result.usage as any).completionTokens).toBeUndefined();
-      });
-    });
-
-    describe('stream', () => {
-      it('should expose usage in stream with AI SDK v5 format', async () => {
-        const model = new MockLanguageModelV2({
-          doStream: async () => {
-            return {
-              stream: convertArrayToReadableStream([
-                { type: 'text-start', id: '1' },
-                { type: 'text-delta', id: '1', delta: 'Hello ' },
-                { type: 'text-delta', id: '1', delta: 'world!' },
-                { type: 'text-end', id: '1' },
-                {
-                  type: 'finish',
-                  finishReason: 'stop',
-                  usage: { inputTokens: 10, outputTokens: 20, totalTokens: 30 },
-                },
-              ]),
-            };
-          },
-        });
-
-        const agent = new Agent({
-          name: 'test-agent',
-          model,
-          instructions: 'You are a helpful assistant',
-        });
-
-        const stream = await agent.stream('Hello');
-
-        // Consume stream to get usage
-        for await (const _ of stream.fullStream) {
-          // Just consume
-        }
-
-        const usage = await stream.usage;
-        console.log('stream usage:', usage);
-
-        // Check that usage exists with v5 format
-        expect(usage).toBeDefined();
-        expect(usage.inputTokens).toBe(10);
-        expect(usage.outputTokens).toBe(20);
-        expect(usage.totalTokens).toBe(30);
-
-        // Ensure backward compatibility keys are NOT present
-        expect((usage as any).promptTokens).toBeUndefined();
-        expect((usage as any).completionTokens).toBeUndefined();
-      });
-    });
-  });
-
-  describe('Agent legacy usage tracking', () => {
-    describe('generateLegacy', () => {
-      it('should expose usage with promptTokens and completionTokens (legacy format)', async () => {
-        // Create a V1 mock that returns usage in legacy format
-        const model = createMockModel({
-          mockText: 'Hello world!',
-        });
-
-        const agent = new Agent({
-          name: 'test-agent',
-          model,
-          instructions: 'You are a helpful assistant',
-        });
-
-        const result = await agent.generateLegacy('Hello');
-
-        // Check that usage exists
-        expect(result.usage).toBeDefined();
-        console.log('generateLegacy usage:', result.usage);
-
-        // Check legacy format keys
-        expect(result.usage.promptTokens).toBe(10);
-        expect(result.usage.completionTokens).toBe(20);
-        expect(result.usage.totalTokens).toBeDefined();
-      });
-    });
-
-    describe('streamLegacy', () => {
-      it('should expose usage with promptTokens and completionTokens (legacy format)', async () => {
-        const model = createMockModel({
-          mockText: 'Hello world!',
-        });
-
-        const agent = new Agent({
-          name: 'test-agent',
-          model,
-          instructions: 'You are a helpful assistant',
-        });
-
-        const result = await agent.streamLegacy('Hello');
-
-        // Consume stream to get usage
-        for await (const _ of result.textStream) {
-          // Just consume
-        }
-
-        const usage = await result.usage;
-        console.log('streamLegacy usage:', usage);
-
-        // Check that usage exists with legacy format
-        expect(usage).toBeDefined();
-        expect(usage.promptTokens).toBeDefined();
-        expect(usage.completionTokens).toBeDefined();
-        expect(usage.totalTokens).toBeDefined();
-        // Legacy format should have promptTokens/completionTokens, not inputTokens/outputTokens
-        expect((usage as any).inputTokens).toBeUndefined();
-        expect((usage as any).outputTokens).toBeUndefined();
-      });
-    });
-
-    describe('generate/stream (currently using legacy implementation)', () => {
-      it('generate should use promptTokens/completionTokens until migration', async () => {
-        const model = createMockModel({
-          mockText: 'Hello world!',
-        });
-
-        const agent = new Agent({
-          name: 'test-agent',
-          model,
-          instructions: 'You are a helpful assistant',
-        });
-
-        const result = await agent.generateLegacy('Hello');
-
-        // Currently using legacy implementation, should have legacy format
-        expect(result.usage).toBeDefined();
-        expect(result.usage.promptTokens).toBe(10);
-        expect(result.usage.completionTokens).toBe(20);
-      });
-
-      it('stream should use promptTokens/completionTokens until migration', async () => {
-        const model = createMockModel({
-          mockText: 'Hello world!',
-        });
-
-        const agent = new Agent({
-          name: 'test-agent',
-          model,
-          instructions: 'You are a helpful assistant',
-        });
-
-        const result = await agent.streamLegacy('Hello');
-
-        // Consume stream
-        for await (const _ of result.textStream) {
-          // Just consume
-        }
-
-        // Currently using legacy implementation, should have legacy format
-        const usage = await result.usage;
-        expect(usage).toBeDefined();
-        expect(usage.promptTokens).toBeDefined();
-        expect(usage.completionTokens).toBeDefined();
-        // Legacy format should have promptTokens/completionTokens, not inputTokens/outputTokens
-        expect((usage as any).inputTokens).toBeUndefined();
-        expect((usage as any).outputTokens).toBeUndefined();
-      });
-    });
-  });
-});
